Декомпозирует компонент Favorites
